Task 1: fix

+++ 0x15-file_io/1-create_file.c
@@ -11,7 +11,7 @@
 int create_file(const char *filename, char *text_content)
 {
 	int file_dp;
-	ssize_t char_wr;
+	int char_wr;
 
 	if (filename == NULL)
 		return (-1);
